Unit-Tests refactored
