+++ src/SortAlgorithms/Testing/InsertionSortingTest.cs
@@ -1,11 +1,12 @@
 namespace Testing
 {
     using SortAlgorithms.SimpleSortings;
+    using Testing.Arrange;
 
     public class InsertionSortingTest
     {
         [Fact]
-        public void Sort_Array_Sorted()
+        public void Sort_ArrayUnique1To10_Sorted()
         {
             var sortingAlgorithm = new InsertionSorting();
             var actualSortedArray = sortingAlgorithm.Sort(RandomNumbersToSort.RandomNumbers_1to10_Unique);
