+++ src/SortAlgorithms/Testing/MergeSortingTest.cs
@@ -1,11 +1,12 @@
 namespace Testing
 {
     using SortAlgorithms.DivideAndConquerSortings;
+    using Testing.Arrange;
 
     public class MergeSortingTest
     {
         [Fact]
-        public void Sort_Array_Sorted()
+        public void Sort_ArrayUnique1To10_Sorted()
         {
             var sortingAlgorithm = new MergeSorting();
             var actualSortedArray = sortingAlgorithm.Sort(RandomNumbersToSort.RandomNumbers_1to10_Unique);
